requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
python-multipart
sqlalchemy
alembic
psycopg2-binary
python-dotenv
